requirements.txt
dotenv
requests
PyMuPDF
fastapi
uvicorn
python-multipart
pytesseract
pdf2image
pillow
opencv-python-headless